backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiohttp==3.9.5
opencv-python-headless==4.10.0.84
numpy==1.26.4
redis==5.0.7
python-dotenv==1.0.1
pydantic==2.8.2